+++ tests/test_doors.py
@@ -1,9 +1,8 @@
 """Test vehicle doors."""
 
 import vehiclepass
+from tests.utils import mock_responses
 from vehiclepass.doors import Doors
-
-from .conftest import mock_responses
 
 
 @mock_responses(status="status/unlocked.json")
@@ -12,19 +11,14 @@ def test_doors_closed(vehicle: vehiclepass.Vehicle) -> None:
     assert isinstance(vehicle.doors, Doors)
     assert vehicle.doors.are_unlocked
     assert vehicle.doors.are_locked is False
-    assert hasattr(vehicle.doors, "unspecified_front")
-    assert hasattr(vehicle.doors, "front_left")
-    assert hasattr(vehicle.doors, "front_right")
-    assert hasattr(vehicle.doors, "rear_left")
-    assert hasattr(vehicle.doors, "rear_right")
-    assert hasattr(vehicle.doors, "tailgate")
-    assert hasattr(vehicle.doors, "inner_tailgate")
-    assert vehicle.doors.front_left == "CLOSED"
-    assert vehicle.doors.front_right == "CLOSED"
-    assert vehicle.doors.rear_left == "CLOSED"  # type: ignore
-    assert vehicle.doors.rear_right == "CLOSED"  # type: ignore
-    assert vehicle.doors.tailgate == "CLOSED"  # type: ignore
-    assert vehicle.doors.inner_tailgate == "CLOSED"  # type: ignore
+    # We call these with getattr to avoid mypy errors, as doors are dynamically
+    # created in the Doors class.
+    assert getattr(vehicle.doors, "front_left") == "CLOSED"
+    assert getattr(vehicle.doors, "front_right") == "CLOSED"
+    assert getattr(vehicle.doors, "rear_left") == "CLOSED"
+    assert getattr(vehicle.doors, "rear_right") == "CLOSED"
+    assert getattr(vehicle.doors, "tailgate") == "CLOSED"
+    assert getattr(vehicle.doors, "inner_tailgate") == "CLOSED"
 
 
 @mock_responses(