+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# v0.4.0
+
+* Added several tests
+* Added TPMS status (`vehicle.tires.system_status`)
+* Refactored tires to include TPMS status for each tire (`vehicle.tires.front_right.status`)
+
+# v0.3.0
+
+* Added `Pressure.bar` unit
+* Fix URLs in pyproject.toml
+* Fix door logic
+
 # v0.3.0
 
 * Added `Pressure.bar` unit